requirements.txt
Django==5.1.2
asgiref==3.8.1
whitenoise==6.7.0
google-generativeai==0.7.2
uvicorn==0.32.0
django-cors-headers==4.4.0
